refactor جدا کردن مشتقات هر کامند و کوئری در فایل های جداگانه
