+++ MakFood.Customer/MakFood.Customer.Application/Commands/Friendship/AcceptFriendship/AcceptFriendshipCommandValidatior.cs
@@ -1,0 +1,12 @@
+﻿using MakFood.Customer.Application.Commands.Friendship.FriendshipOperationBase;
+
+namespace MakFood.Customer.Application.Commands.Friendship.AcceptFriendship
+{
+    public class AcceptFriendshipCommandValidatior : FriendshipOperationBaseCommandValidation
+    {
+        public AcceptFriendshipCommandValidatior()
+        {
+            
+        }
+    }
+}
