+++ MakFood.Customer/MakFood.Customer.Application/Commands/Friendship/CreateFriendship/CreateFriendshipCommandRespone.cs
@@ -1,0 +1,11 @@
+﻿namespace MakFood.Customer.Application.Commands.Friendship.CreateFriendship
+{
+    public class CreateFriendshipCommandRespone
+    {
+        public Guid Id { get; set; }
+    }
+
+    
+}
+
+
