+++ MakFood.Customer/MakFood.Customer.Application/Commands/Friendship/RejectFriendship/RejectFriendshipCommandValidator.cs
@@ -1,0 +1,12 @@
+﻿using MakFood.Customer.Application.Commands.Friendship.FriendshipOperationBase;
+
+namespace MakFood.Customer.Application.Commands.Friendship.RejectFriendship
+{
+    public class RejectFriendshipCommandValidator : FriendshipOperationBaseCommandValidation
+    {
+        public RejectFriendshipCommandValidator()
+        {
+            
+        }
+    }
+}
