+++ MakFood.Customer/MakFood.Customer.Application/Commands/Friendship/RevokeFriendship/RevokeFriendshipCommandValidator.cs
@@ -1,0 +1,11 @@
+﻿using MakFood.Customer.Application.Commands.Friendship.FriendshipOperationBase;
+
+namespace MakFood.Customer.Application.Commands.Friendship.RevokeFriendship
+{
+    public class RevokeFriendshipCommandValidator : FriendshipOperationBaseCommandValidation
+    {
+        public RevokeFriendshipCommandValidator()
+        {
+        }
+    }
+}
